Refactor code-sharing between URL AST and plain web.py Catalogs handler.
1. Split part of ermrest_apis into apicore module to allow reuse of
   basic request_init without circular module dependencies

2. Define @web_method() decorator to allow shared wrapping with
   request_init/final, logging, and core exception to REST exception
   mappings whether the handler is invoked from our URL parsing
   Dispatcher class or directly by web.py

3. Remove useless inheritence of Api class by Catalogs handler, to
   avoid an __init__() time depdendency on web.ctx content that isn't
   set until the web_method() wrapper runs.

4. Try to consolidate some of the logging/tracing logic to reduce
   duplicate code and odd variants.  This may need further work.
