+++ ermrest/makefile-vars
@@ -5,6 +5,7 @@ ERMREST_SHARE_FILES= \
 
 ERMREST_PYTHON_FILES= \
 	__init__.py \
+	apicore.py \
 	ermrest_apis.py \
 	ermrest.wsgi \
 	sanepg2.py \
